+++ src/main/kotlin/org/bytecamp/program_repair/astor_plugin/actions/AstorPrintArgumentAction.kt
@@ -6,7 +6,7 @@ import com.intellij.openapi.components.service
 import com.intellij.openapi.ui.Messages
 import org.bytecamp.program_repair.astor_plugin.services.AstorProjectService
 
-class AstorHelloAction: AnAction() {
+class AstorPrintArgumentAction: AnAction() {
     override fun actionPerformed(e: AnActionEvent) {
         val service = e.project?.service<AstorProjectService>()!!
         Messages.showMessageDialog(e.project, "To execute astor with arguments \n" + service.getConfig().toArgs().joinToString("\n"), "Astor", null)