+++ src/main/kotlin/org/bytecamp/program_repair/astor_plugin/services/AstorProjectService.kt
@@ -10,6 +10,7 @@ import com.intellij.openapi.ui.Messages
 import com.intellij.openapi.vfs.VirtualFile
 import fr.inria.main.evolution.AstorMain
 import java.io.IOException
+import kotlin.math.log
 
 
 object Keys {
@@ -24,7 +25,7 @@ object Keys {
 
 class AstorConfig {
     var mode: String = "jGenProg"
-    var pkg: String = "com.company"
+    var pkg: String = ""
     var location: String = ""
     var src: String = ""
     var srcTest: String = ""
@@ -48,6 +49,7 @@ class AstorConfig {
             binTest
         )
     }
+
     override fun toString(): String {
         return toArgs().joinToString(" ")
     }
@@ -56,42 +58,74 @@ class AstorConfig {
 @Service
 class AstorProjectService(val project: Project) {
     val main: AstorMain = AstorMain()
+    val logger = com.intellij.openapi.diagnostic.Logger.getInstance("AstorProjectService")
 
     fun getCommonPackage(file: VirtualFile, prefix: String): String {
         return if (file.isDirectory) {
             if (file.children.size > 1) {
-                prefix + "." + file.name
+                if (prefix.isEmpty()) {
+                    file.name
+                } else {
+                    prefix + "." + file.name
+                }
             } else if (file.children.isEmpty()) {
                 prefix
             } else {
-                getCommonPackage(file, prefix + "." + file.name)
+                if (file.children[0].name == "java") {
+                    getCommonPackage(file.children[0], prefix)
+                } else {
+                    getCommonPackage(file.children[0], prefix + "." + file.name)
+                }
             }
         } else {
-            throw IOException("Cannot getCommonPackage in file " + file.path)
+            prefix
         }
+    }
+    fun getBuildSystem(): String {
+        val base = project.projectFile?.parent?.parent
+        val system =  if (base?.findChild("build.gradle") != null || base?.findChild("build.gradle.kts") != null)  {
+            "gradle"
+        } else if (base?.findChild("pom.xml") != null) {
+            "maven"
+        } else {
+            "others"
+        }
+        logger.info("Base " + base?.path + " build system " + system)
+        return system
     }
 
     fun getConfig(): AstorConfig {
         val config = AstorConfig()
         config.location = project.basePath!!
-        for (module in ModuleManager.getInstance(project).modules) {
-            val srcRoots = ModuleRootManager.getInstance(module).sourceRoots
-            if (srcRoots.size != 1) {
-                Messages.showWarningDialog(project, "Can only process one source root", "Astor Warn")
-            }
-            if (module.name == "main") {
-                config.src = srcRoots[0].path
-                config.pkg = getCommonPackage(srcRoots[0], "")
-            } else if (module.name == "test") {
-                config.srcTest = srcRoots[0].path
-            }
+        val modules = ModuleManager.getInstance(project).modules
 
+        for (module in modules) {
+            logger.info("Module %s path %s".format(module.name, module.moduleFilePath))
+            for (srcRoot in ModuleRootManager.getInstance(module).sourceRoots) {
+                logger.info("Source root " + srcRoot.path)
+                if (srcRoot.name == "java") {
+                    if (srcRoot.path.contains("main")) {
+                        config.src = srcRoot.path
+                        config.pkg = getCommonPackage(srcRoot, "")
+                    } else if (srcRoot.path.contains("test")) {
+                        config.srcTest = srcRoot.path
+                    }
+                }
+            }
         }
-        val compiler = CompilerProjectExtension.getInstance(project)!!
-
-        config.bin =  compiler.compilerOutput?.path!!
-        config.binTest = compiler.compilerOutput?.path!!
-
+        when (getBuildSystem()) {
+            "gradle" -> {
+                config.bin = project.projectFile!!.parent.path + "build/classes/java/main"
+                config.binTest = project.projectFile!!.parent.path + "build/classes/java/test"
+            }
+            "maven" -> {
+                config.bin = project.projectFile!!.parent.path + "target/classes"
+                config.binTest = project.projectFile!!.parent.path + "target/test-classes"
+            }
+            else -> {
+                Messages.showErrorDialog(project, "Does not support build systems other than gradle or maven", "Astor")
+            }
+        }
         return config
     }